add odos v3 volume and fees (#3753)
